refactor: keep only index test and styles at a component folder root
