+++ FateConnect/Web/src/pages/Home/components/LandingDescription/constants/index.ts
