+++ FateConnect/Web/src/pages/Home/components/LandingHowItWorks/constants/index.ts
