+++ FateConnect/Web/src/pages/Home/components/LandingLoginCard/constants/index.ts
