+++ FateConnect/Web/src/pages/Home/components/LandingLoginCard/schema/index.ts
